Added much better support for typing on a keyboard
This has been put off in the past, but the most practical use case
seems to be speaking alongside typing. This commit adds supports
for most important special keys needed for typing. There is a
lingering issue with how to type uppercase letters which currently
has no plans of being resolved.