interpreter: Gracefully deal with embedded punctuation
Now the interpreter will perform a grammar look up on each word
and punctuation mark separately. For example, we would expect the
following grammar lookups to occur:
  "foo, bar" -> grammar["foo"], grammar[","], grammar["bar"]

Previously, it would instead behave as follows:
  "foo, bar" -> grammar["foo,"], grammar["bar"]

This caused a lot of false negatives and made punctuation a pain
to speak.

Now consider the case where "foo" is a grammar function. This
implements the following behavior:
  "foo, bar. baz" -> foo(", bar. baz")

Notice that for grammar functions the punctuation of the remaining
text is left intact. This makes it easier to speak punctuation when
narrating large blocks of prose.

Things got a little complicated because of several edge cases,
mostly relating to words beginning with punctuation or spaces.