interpreter: Load grammars on start up
Previously the grammars were precompiled into `grammar_compiled`
files. This approach was used because I originally thought it
would help start up time. It turns out start up time is so fast
that it doesn't matter.

Now the grammar files are parsed when the interpreter starts. This
removes a lot of complexity and will help when simplifying the
format grammars are written in. It will also make live reloading the
grammars easier.