grammars: Change directory structure
Previously each grammar was a module in its own directory. This
resulted from the previous approach of precompiling the grammars.
In specific, grammar functions were split into a separate file.

Now that the grammars are loaded at start up, grammar functions
can just be stored in the grammar file directly. This eliminates
the need for the grammar subdirectories. Instead, now each grammar
is just a file in the grammars directory.